Refactoring file structure
- The file structure was all jacked up so I am reorganzied into a standard file setup.
