+++ src/app/blocks/CenteredBlock.jsx
