+++ src/app/blocks/ImageBlock.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { BaseBlock } from 'components/base/blocks';
+import { BaseBlock } from 'app/blocks';
 
 const StyledBlock = styled(BaseBlock)`
   object-fit: cover;
