+++ src/app/blocks/index.js
@@ -1,0 +1,5 @@
+import BaseBlock from 'app/blocks/BaseBlock';
+import CenteredBlock from 'app/blocks/CenteredBlock';
+import ImageBlock from 'app/blocks/ImageBlock';
+
+export { BaseBlock, CenteredBlock, ImageBlock };
