--- src/components/base/blocks/index.js
@@ -1,5 +1,0 @@
-import BaseBlock from 'components/base/blocks/BaseBlock';
-import CenteredBlock from 'components/base/blocks/CenteredBlock';
-import ImageBlock from 'components/base/blocks/ImageBlock';
-
-export { BaseBlock, CenteredBlock, ImageBlock };
