+++ static/admin/config.yml
@@ -17,21 +17,21 @@ collections:
           - {label: "Template Key", name: "templateKey", widget: "hidden"}
           - {label: "First Name", name: "firstName", widget: "string"}
           - {label: "Last Name", name: "lastName", widget: "string"}
-      - file: "src/pages/resume/index.md"
-        label: "Resume"
-        name: "resume"
-        slug: "resume"
-        fields:
-          - {label: "Template Key", name: "templateKey", widget: "hidden"}
-          - {label: "Title", name: "title", widget: "string"}
-          - {label: "Intro", name: "intro", widget: "markdown"}
-          - {label: "Job Title", name: "jobTitle", widget: list, fields: [{label: Company, name: company, widget: string}, {label: Positon, name: position, widget: string}, {label: "Employment Dates", name: employmentDates, widget: string}, {label: Description, name: description, widget: string}]}
-          - {label: "Education", name: "education", widget: list, fields: [{label: School, name: school, widget: string}, {label: Degree, name: degree, widget: string}]}
-      - file: "src/pages/about/index.md"
-        label: "About"
-        name: 'about'
-        slug: 'about'
-        fields:
-          - {label: "Template Key", name: "templateKey", widget: "hidden"}
-          - {label: "Name", name: "name", widget: "string"}
-          - {label: "Bio", name: "bio", widget: "markdown"}
+      # - file: "src/pages/resume/index.md"
+      #   label: "Resume"
+      #   name: "resume"
+      #   slug: "resume"
+      #   fields:
+      #     - {label: "Template Key", name: "templateKey", widget: "hidden"}
+      #     - {label: "Title", name: "title", widget: "string"}
+      #     - {label: "Intro", name: "intro", widget: "markdown"}
+      #     - {label: "Job Title", name: "jobTitle", widget: list, fields: [{label: Company, name: company, widget: string}, {label: Positon, name: position, widget: string}, {label: "Employment Dates", name: employmentDates, widget: string}, {label: Description, name: description, widget: string}]}
+      #     - {label: "Education", name: "education", widget: list, fields: [{label: School, name: school, widget: string}, {label: Degree, name: degree, widget: string}]}
+      # - file: "src/pages/about/index.md"
+      #   label: "About"
+      #   name: 'about'
+      #   slug: 'about'
+      #   fields:
+      #     - {label: "Template Key", name: "templateKey", widget: "hidden"}
+      #     - {label: "Name", name: "name", widget: "string"}
+      #     - {label: "Bio", name: "bio", widget: "markdown"}
